+++ src/main/kotlin/no/skatteetaten/aurora/mokey/model/ApplicationData.kt
@@ -1,10 +1,6 @@
 package no.skatteetaten.aurora.mokey.model
 
 import com.fasterxml.jackson.databind.JsonNode
-import no.skatteetaten.aurora.mokey.service.Endpoint
-import no.skatteetaten.aurora.mokey.service.HealthResponse
-import no.skatteetaten.aurora.mokey.service.InfoResponse
-import no.skatteetaten.aurora.mokey.service.ManagementLinks
 import no.skatteetaten.aurora.utils.Either
 
 data class ApplicationData(
